requirements.txt
python-dotenv
gtts
pydub
elevenlabs
